Add swagger, minor bug fixed

+++ apps/marks/models/mastery.py
@@ -5,7 +5,7 @@ from mixins.models import DateTimeMixin
 
 class Mastery(DateTimeMixin):
     tank = models.OneToOneField('directory.Tank', on_delete=models.CASCADE, related_name='mastery')
-    class_3 = models.IntegerField()
-    class_2 = models.IntegerField()
-    class_1 = models.IntegerField()
-    master = models.IntegerField()
+    class_3 = models.PositiveSmallIntegerField()
+    class_2 = models.PositiveSmallIntegerField()
+    class_1 = models.PositiveSmallIntegerField()
+    master = models.PositiveSmallIntegerField()
